pyproject.toml
[project]
name = "og-agents"
version = "0.1.0"
description = ""
authors = [
    {name = "Kolya Nosenko",email = "nosenko.nicholas@gmail.com"}
]
readme = "README.md"
requires-python = "=3.13.5"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain (>=1.2.0,<2.0.0)",
    "langchain-core (>=1.2.5,<2.0.0)",
    "langgraph (>=1.0.5,<2.0.0)",
    "langchain-openai (>=1.1.6,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "rdflib (>=7.5.0,<8.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-huggingface (>=1.2.0,<2.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "transformers (>=4.57.3,<5.0.0)",
    "accelerate (>=1.12.0,<2.0.0)",
    "pyvis (>=0.3.2,<0.4.0)",
    "networkx (>=3.6.1,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "owlready2 (>=0.49,<0.50)",
    "streamlit (>=1.52.2,<2.0.0)",
]

[tool.poetry]
packages = [{include = "og_agents", from = "src"}]

[tool.poetry.group.test.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pylint = "^3.3.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.isort]
profile = 'black'
line_length = 88
src_paths = ["src", "tests"]
known_first_party = ["knowledge_rag"]

[tool.pylint.main]
py-version = "3.13"
jobs = 0                    # 0 = авто-визначення за кількістю ядер
recursive = true


[tool.pylint.message_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    # "R0903", too-few-public-methods
]

[tool.pylint.format]
max-line-length = "88"
